Symlink Locks folder only if NFS session

+++ install.sh
@@ -26,18 +26,21 @@ export HOMEBREW_TEMP=/tmp/\$USER/Homebrew/Temp
 mkdir -p \$HOMEBREW_CACHE
 mkdir -p \$HOMEBREW_TEMP
 
-HOMEBREW_LOCKS_TARGET=/tmp/\$USER/Homebrew/Locks
-HOMEBREW_LOCKS_FOLDER=\$HOME/.brew/var/homebrew
+if df -T autofs,nfs \$HOME 1>/dev/null
+then
+  HOMEBREW_LOCKS_TARGET=/tmp/\$USER/Homebrew/Locks
+  HOMEBREW_LOCKS_FOLDER=\$HOME/.brew/var/homebrew
 
-mkdir -p \$HOMEBREW_LOCKS_TARGET
-mkdir -p \$HOMEBREW_LOCKS_FOLDER
+  mkdir -p \$HOMEBREW_LOCKS_TARGET
+  mkdir -p \$HOMEBREW_LOCKS_FOLDER
 
-# Symlink to Locks target folders
-# If not already a symlink
-if ! [[ -L \$HOMEBREW_LOCKS_FOLDER && -d \$HOMEBREW_LOCKS_FOLDER ]]; then
-  echo "Creating symlink for Locks folder"
-  rm -rf \$HOMEBREW_LOCKS_FOLDER
-  ln -s \$HOMEBREW_LOCKS_TARGET \$HOMEBREW_LOCKS_FOLDER
+  # Symlink to Locks target folders
+  # If not already a symlink
+  if ! [[ -L \$HOMEBREW_LOCKS_FOLDER && -d \$HOMEBREW_LOCKS_FOLDER ]]; then
+     echo "Creating symlink for Locks folder"
+     rm -rf \$HOMEBREW_LOCKS_FOLDER
+     ln -s \$HOMEBREW_LOCKS_TARGET \$HOMEBREW_LOCKS_FOLDER
+  fi
 fi
 EOL
